Kapt tests: minor, fix path escaping on Windows
Path separators are replaced in classpath arguments from ':' to ';' on
Windows (see `preprocessPathSeparators`), so we should escape the
argument in case it contains ';' as well. Otherwise it leads to kapt4
tests failing in subsequent commits, where command lines are a bit
different because apt and compilation are invoked separately in kapt4.